Split S4 into two files to make it more readable
